Property feature dictionary

This function returns a dictionary of searchable property features. These features can be used as search parameters in the count_listings, listings, listing_counts and map_listings functions.

dictionary_listing_features()

Value

a data_frame

See also

Other dictionary: dictionary_property_types

Other listing search: dictionary_property_types, listing_counts, listings, map_listings

Examples

# NOT RUN {
dictionary_listing_features()
# }